+++ Dockerfile
@@ -1,0 +1,7 @@
+FROM node:alpine
+
+WORKDIR /server
+COPY package*.json ./
+
+RUN npm install
+CMD ["npm","start"]
