+++ db/Dockerfile
@@ -1,0 +1,2 @@
+FROM postgres:16.3
+COPY build.sql /docker-entrypoint-initdb.d/
