+++ server/bot/bot.js
