+++ server/bot/commands/addMemberToTeam.js
