+++ server/bot/commands/createTeam.js
