+++ server/bot/commands/getInfoMeeting.js
