+++ server/bot/commands/getMeetings.js
