+++ server/bot/commands/setCommands.js
