+++ server/bot/commands/start.js
