+++ server/bot/config.js
